fix : post 도메인 내에  매핑테이블 mapping 폴더로 이동
