+++ src/main/java/com/vinny/backend/post/domain/Post.java
@@ -2,6 +2,10 @@ package com.vinny.backend.post.domain;
 
 import com.vinny.backend.User.domain.User;
 import com.vinny.backend.common.domain.BaseEntity;
+import com.vinny.backend.post.domain.mapping.PostBrandHashtag;
+import com.vinny.backend.post.domain.mapping.PostShopHashtag;
+import com.vinny.backend.post.domain.mapping.PostStyleHashtag;
+import com.vinny.backend.post.domain.mapping.UserPost;
 import jakarta.persistence.*;
 import lombok.*;
