+++ src/main/java/com/vinny/backend/post/domain/mapping/UserPost.java
@@ -1,7 +1,8 @@
-package com.vinny.backend.post.domain;
+package com.vinny.backend.post.domain.mapping;
 
 import com.vinny.backend.User.domain.User;
 import com.vinny.backend.common.domain.BaseEntity;
+import com.vinny.backend.post.domain.Post;
 import jakarta.persistence.*;
 import lombok.*;
