+++ src/main/java/com/vinny/backend/post/dto/PostResponseDto.java
@@ -1,0 +1,17 @@
+package com.vinny.backend.post.dto;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+public record PostResponseDto(
+        Long postId,
+        AuthorDto author,
+        String content,
+        List<String> images,
+        LocalDateTime createdAt,
+        int likesCount,
+        boolean isLikedByMe,
+        ShopDto shop,
+        StyleDto style,
+        BrandDto brand
+) {}
